==> 01-preparation.ipynb <==
# SQL und statistische Analyse

import pandas as pd
import sqlite3

sql = sqlite3.connect("../newsticker.db")

docs = pd.read_sql("SELECT * FROM news", sql)

docs["time"].max()

docs = docs[(docs["time"]>='2022-01-01')]
docs

docs["quality"] = "uncertain"
docs.loc[docs["comments"] < 10, "quality"] = "bad"
docs.loc[docs["comments"] > 50, "quality"] = "good"

docs.value_counts("quality").plot.bar()

len(docs)

docs

# docs[docs["quality"]!="uncertain"].set_index("id")[["headline", "comments", "time", "quality"]].to_csv("newsticker-2022-2023-good-bad.csv.gz")
