==> 23-similarity-cross-encoder.ipynb <==
# Using a cross encoder to optimize scoring

As you have seen in the [previous notebook](22-similarity-rrf.ipynb), reciprocal
rank fusion can be used to combine multiple result lists into one. As the
scores are incommensurable, only the rank was considered for this.

However, we can improve on that by using a [cross encoder](https://sbert.net/examples/cross_encoder/applications/README.html).
The cross encoder is a (modified) BERT model which checks the similarity of two
sentences or how well an answer fits to a question. By using the last layer of
the BERT transformer network, a probability can be extracted and will be used as
a score.

There are many cross encoders (sometimes also called rerankers) available. You can find a list on 
[Hugging Face](https://huggingface.co/models?pipeline_tag=text-ranking). Be careful,
not all have the same calling conventions and the best models are often quite slow
and mostly suitable for offline use.

## Load data (from previous notebook)

import json
with open("sentences.json") as f:
    sentences = json.load(f)

len(sentences)

import numpy as np
with open("sentences-mqa.npy", "rb") as f:
    sembeddings = np.load(f)

## Retrieval

import numpy as np
import pandas as pd
def search(query, text, corpus_embeddings, bi_encoder, cross_encoder, query_prompt_name=None, top_k=100):
    # code query to restrict search space
    question_embedding = bi_encoder.encode(query, normalize_embeddings=True, prompt_name=query_prompt_name)
    
    # Determine similarity (vectors are normalized)
    sim = model.similarity(question_embedding, corpus_embeddings)[0].numpy() 
    
    # Get most similar top_k by sorting
    hits = [ { "id": i, "text": text[i], "score": sim[i] } 
                     for i in sim.argsort()[::-1][0:top_k] ]

    # Consider only top hits for re-ranking
    cross_input = [[query, hit["text"]] for hit in hits]
    # cross-encode (this takes most time)
    cross_scores = cross_encoder.predict(cross_input)

    # Integrate cross-scores in original hits (this would be easier with pandas)
    for i in range(len(cross_scores)):
        hits[i]["cross-score"] = cross_scores[i]

    # nre-sort by cross-score, descending!
    hits = sorted(hits, key=lambda x: x["cross-score"], reverse=True)
    
    # Return top-20 results of re-ranker as dataframe
    return pd.DataFrame(hits[0:20]).set_index("id")

# bi-encoder is needed
from sentence_transformers import SentenceTransformer
model = SentenceTransformer('multi-qa-MiniLM-L6-cos-v1')

# cross encoder
from sentence_transformers import CrossEncoder, util
cross_encoder = CrossEncoder('cross-encoder/ms-marco-MiniLM-L-6-v2')

pd.set_option('display.max_colwidth', 0)

search("Is the climate crisis worse in poorer countries?", sentences, sembeddings, model, cross_encoder).style.background_gradient(cmap='coolwarm')

# try with different model
# if you believe the cross encoder, the document with the lowest cross encoding
# score might be a measure of the quality of the embedding model
model2 = SentenceTransformer('google/embeddinggemma-300m')
with open("sentences-gemma.npy", "rb") as f:
    sembeddings2 = np.load(f)

search("Is the climate crisis worse in poorer countries?", 
       sentences, sembeddings2, model2, cross_encoder, query_prompt_name="query").style.background_gradient(cmap='coolwarm')

model3 = SentenceTransformer("Snowflake/snowflake-arctic-embed-l-v2.0", trust_remote_code=True)
with open("sentences-arctic.npy", "rb") as f:
    sembeddings3 = np.load(f)

search("Is the climate crisis worse in poorer countries?", 
       sentences, sembeddings3, model3, cross_encoder, query_prompt_name="query").style.background_gradient(cmap='coolwarm')

## Alternative cross encoder

cross_encoder2 = CrossEncoder('mixedbread-ai/mxbai-rerank-large-v1')

search("Is the climate crisis worse in poorer countries?", sentences, sembeddings, model,
       cross_encoder2).style.background_gradient(cmap='coolwarm')

search("Is the climate crisis worse in poorer countries?", sentences, sembeddings2, model2, 
       cross_encoder2, query_prompt_name="query").style.background_gradient(cmap='coolwarm')

search("Is the climate crisis worse in poorer countries?", sentences, sembeddings3, model3, 
       cross_encoder2, query_prompt_name="query").style.background_gradient(cmap='coolwarm')
