==> 31-sbert-finetune.ipynb <==
# Finetuning a sentence similarity model

[SBERT](https://sbert.net) offers a framework for calculating similarities between sentences (of course, also larget entities can be used due to the increased context length). A lot of fantastic models already exist and can be found on the [MTEB](https://huggingface.co/spaces/mteb/leaderboard).

To show how well finetuning works, we deliberately start with an unsuitable model, again using ModernBERT. We then find out that after finetuning our results improve considerably.

# Load the dataset

from datasets import load_dataset

train_dataset = load_dataset("sentence-transformers/all-nli", "pair-score", split="train")

Inspect

train_dataset

First row

train_dataset[0]

Looks strange, take a deeper look

import pandas as pd
pd.set_option('display.max_colwidth', None)
df = train_dataset.to_pandas()
df.head()

How many different scores are present?

df["score"].value_counts()

Take a look at the sample!

df.sample(20, random_state=42)

# Create a SBERT model based on ModernBERT

from sentence_transformers import SentenceTransformer

model_name = "answerdotai/ModernBERT-base"
model = SentenceTransformer(model_name, device="cuda")

model.to('cuda')

We need a loss function which tells us how our models rates the sentences compared to the ground truth.

from sentence_transformers.losses import CoSENTLoss
loss = CoSENTLoss(model)

SBERT has a training (i.e. finetuning) framework integrated

from sentence_transformers import SentenceTransformerTrainingArguments
from sentence_transformers.training_args import BatchSamplers
args = SentenceTransformerTrainingArguments(
    # Required parameter:
    output_dir="models/ModernSBERT",
    # Optional training parameters:
    num_train_epochs=1,
    per_device_train_batch_size=16,
    per_device_eval_batch_size=16,
    learning_rate=2e-5,
    warmup_ratio=0.1,
    fp16=False,
    bf16=True,
    batch_sampler=BatchSamplers.NO_DUPLICATES,  # losses that use "in-batch negatives" benefit from no duplicates
)

We need to evaluate the similarity for the different parts of the dataset (training, validation, test)

from sentence_transformers.evaluation import EmbeddingSimilarityEvaluator, SimilarityFunction

# Load the STSB dataset (https://huggingface.co/datasets/sentence-transformers/stsb)
eval_dataset = load_dataset("sentence-transformers/all-nli", "pair-score", split="dev")

# Initialize the evaluator
dev_evaluator = EmbeddingSimilarityEvaluator(
    sentences1=eval_dataset["sentence1"],
    sentences2=eval_dataset["sentence2"],
    scores=eval_dataset["score"],
    main_similarity=SimilarityFunction.COSINE,
    name="all-nli-dev",
)

test_dataset = load_dataset("sentence-transformers/all-nli", "pair-score", split="test")

test_evaluator = EmbeddingSimilarityEvaluator(
    sentences1=test_dataset["sentence1"],
    sentences2=test_dataset["sentence2"],
    scores=test_dataset["score"],
    main_similarity=SimilarityFunction.COSINE,
)

Check how well the model performs on the test dataset

test_evaluator(model)

Run the training (takes too long in the live course)

from sentence_transformers import SentenceTransformerTrainer
trainer = SentenceTransformerTrainer(
    model=model,
    args=args,
    train_dataset=train_dataset,
    eval_dataset=eval_dataset,
    loss=loss,
    evaluator=dev_evaluator,
)
trainer.train()

model.save_pretrained("ModernSBERT-base-all-nli/final")

Run the evaluator again on the finetuned dataset

model = SentenceTransformer("./ModernSBERT-base-all-nli/final")

test_evaluator(model)
