==> 32-create-sbert-data-qwen-reranker.ipynb <==
# Improve the dataset used for SBERT finetuning

The dataset used for finetuning the ModernBERT model for similarity looks strange with respect to the very opinionated similarity values. We can do better by using a (very good) reranking model and calculating the similarities with this.

It turns out that these also improves the performance of the resulting finetuned model!

import torch
from transformers import AutoModel, AutoTokenizer, AutoModelForCausalLM

This is an instruction following model and needs to be prompted

def format_instruction(instruction, query, doc):
    output = "<Instruct>: {instruction}\n<Query>: {query}\n<Document>: {doc}".format(instruction=instruction,query=query, doc=doc)
    return output

def process_inputs(pairs):
    inputs = tokenizer(
        pairs, padding=False, truncation='longest_first',
        return_attention_mask=False, max_length=max_length - len(prefix_tokens) - len(suffix_tokens)
    )
    for i, ele in enumerate(inputs['input_ids']):
        inputs['input_ids'][i] = prefix_tokens + ele + suffix_tokens
    inputs = tokenizer.pad(inputs, padding=True, return_tensors="pt", max_length=max_length)
    for key in inputs:
        inputs[key] = inputs[key].to(model.device)
    return inputs

We are interested in the probabilities of yet and no, therefore we take a look at the last layer and get these:

@torch.no_grad()
def compute_logits(inputs, **kwargs):
    batch_scores = model(**inputs).logits[:, -1, :]
    true_vector = batch_scores[:, token_true_id]
    false_vector = batch_scores[:, token_false_id]
    batch_scores = torch.stack([false_vector, true_vector], dim=1)
    batch_scores = torch.nn.functional.log_softmax(batch_scores, dim=1)
    scores = batch_scores[:, 1].exp().tolist()
    return scores

# this is large and slow, but slightly better:
# model_name = "Qwen/Qwen3-Reranker-4B"
model_name = "Qwen/Qwen3-Reranker-0.6B"
tokenizer = AutoTokenizer.from_pretrained(model_name, padding_side='left')
model = model = AutoModelForCausalLM.from_pretrained(model_name, dtype=torch.float16, attn_implementation="flash_attention_2").cuda().eval()

Use the *flawed* dataset again

from datasets import load_dataset
train_dataset = load_dataset("sentence-transformers/all-nli", "pair-score", split="train")

df = train_dataset.to_pandas()

import pandas as pd
pd.set_option('display.max_colwidth', None)

df.head(20).style.background_gradient(cmap='coolwarm')

Get the tokens ids for `yes` and `no`, used for calculating the probabilities later

token_false_id = tokenizer.convert_tokens_to_ids("no")
token_true_id = tokenizer.convert_tokens_to_ids("yes")

We use the Qwen chat template (more about that later) and a special system prompt:

max_length = 8192

prefix = "<|im_start|>system\nJudge whether the documents are similar . Note that the answer can only be \"yes\" or \"no\".<|im_end|>\n<|im_start|>user\n"
suffix = "<|im_end|>\n<|im_start|>assistant\n<think>\n\n</think>\n\n"
prefix_tokens = tokenizer.encode(prefix, add_special_tokens=False)
suffix_tokens = tokenizer.encode(suffix, add_special_tokens=False)

Create a temporary dataframe containing the start of the dataset and some arbitrary samples:

dft = pd.concat([df.head(20), df.sample(20, random_state=42)])

Create the instruct codes for prompting the LLM:

pairs = [format_instruction("Given two sentences, calculate their similarity", query, doc) 
           for query, doc in zip(dft["sentence1"], dft["sentence2"])]

Run the actual reranking process:

%%time
# Tokenize the input texts
max_length = 8192
inputs = process_inputs(pairs)
scores = compute_logits(inputs)

# integrate scores
dft["scores"] = scores

dft.style.background_gradient(cmap='coolwarm')
