==> 43-vllm-direct.ipynb <==
# Text generation with `vllm` as a library

[`vllm`](https://github.com/vllm-project/vllm) can be used both as a library and
a server. In this notebook, we focus on the library. Compared to `transformes`,
a lot of the complexity has been abstracted away.

from vllm import LLM, SamplingParams

This can take a while, as the model is loaded and optimized:

model_name = "Qwen/Qwen3-8B"
llm = LLM(model=model_name, max_model_len=16384, trust_remote_code=True)

The `messages` are exactly the same:

messages = [
    {"role": "system", "content": "Du bist ein hilfreicher Assistent."},
    {"role": "user", "content": "Erkläre den Heise Verlag!"}
]

Sampling parameters are passed separately.

sampling_params = SamplingParams(
  max_tokens=1024,
  temperature=0.0,
)

%%time
output = llm.chat(messages=messages, sampling_params=sampling_params)

for o in output:
    prompt = o.prompt
    generated_text = o.outputs[0].text
    print(f"Prompt: {prompt!r}, Generated text: {generated_text!r}")
    print(generated_text)

Avoid the thinking phase:

output = llm.chat(messages=messages, sampling_params=sampling_params, 
                  chat_template_kwargs={"enable_thinking": False})

print(output[0].outputs[0].text)
