==> 52b-llama32-1.ipynb <==
# Companion notebook for finetune of Llama 3.2 with 1.7 billion parameters

Evaluation of the finetune with specific prompts about LLMs and GRPO.

import torch
from transformers import AutoModelForCausalLM, AutoTokenizer

model_name = 'meta-llama/Llama-3.2-1B-Instruct'
dataset_name = 'llm-abstract-dataset.jsonl.xz'
output_dir = "runs/" + model_name.split("/")[-1] + "-" + dataset_name.split(".")[0]

model = AutoModelForCausalLM.from_pretrained(model_name, 
                                             revision='main',
                                             trust_remote_code=False,
                                             attn_implementation='flash_attention_2',
                                             dtype=torch.bfloat16,
                                             use_cache=False,
                                             device_map="cuda",
                                             low_cpu_mem_usage=True)

from peft import PeftModel
model = PeftModel.from_pretrained(model, output_dir)

model = model.merge_and_unload()

tokenizer = AutoTokenizer.from_pretrained(model_name)

model.eval()

print(tokenizer.chat_template)

messages = [{ "content": """You are an educated researcher and always answer in correct scientific terms.
                             You are very deep into LLM and its methodologies. You are very creative.""", 
              "role": "system" }, 
            { "content": "Write an abstract with the title 'New Training Methods for LLMs'", 
              "role": "user" },
            ]

inputs = tokenizer.apply_chat_template(messages, add_generation_prompt=True, return_tensors = "pt").to("cuda")

print(tokenizer.batch_decode(inputs)[0])

outputs = model.generate(inputs, max_new_tokens = 512, use_cache = True,
                         do_sample=True, temperature=0.7, top_k=25, top_p=0.8)

print(tokenizer.batch_decode(outputs)[0])

outputs = model.generate(inputs, max_new_tokens = 512, use_cache = True,
                         do_sample=True, temperature=0.7, top_k=25, top_p=0.8)

print(tokenizer.batch_decode(outputs)[0])

outputs = model.generate(inputs, max_new_tokens = 512, use_cache = True,
                         do_sample=False, temperature=0)

print(tokenizer.batch_decode(outputs)[0])

outputs = model.generate(inputs, max_new_tokens = 512, use_cache = True,
                         do_sample=False, temperature=0)
print(tokenizer.batch_decode(outputs)[0])

messages = [{ "content": """You are an educated researcher and always answer in correct scientific terms.
                             You are very deep into AI and its methodologies. You are very creative.""", 
              "role": "system" }, 
            { "content": "Write an abstract with the title 'Frontiers in GRPO Finetuning'", 
              "role": "user" },
            ]
